Pipeline using shell script
