+++ AIFI_ws/src/sensor_config_edit/launch/sensor_config_edit.launch
@@ -5,5 +5,6 @@
   <arg name='timeout' default='10'/>
   <arg name='agent' default='agent'/>
   <node pkg="sensor_config_edit" type="create_sensor_config.py" name="FIT_test" output='screen'/>
+  <node pkg="sensor_config_edit" type="scenario_manager.py" name="scenario_Manager" output='screen'/>
 
 </launch>
