+++ AIFI_ws/src/sensor_config_edit/src/automation.sh
@@ -1,0 +1,12 @@
+#!/bin/bash
+source /home/autoware/.bashrc
+source /opt/ros/melodic/setup.bash
+source /home/autoware/AIFI_ws/devel/setup.bash
+
+gnome-terminal -e "sshpass -p hmc2020 ssh -t hmcl@192.168.0.121 './FIT_automation.sh'"
+sleep 1
+gnome-terminal -e "roslaunch carla_autoware_agent carla_autoware_agent.launch"
+# sleep 10
+# gnome-terminal -e "roslaunch sensor_config_edit sensor_config_edit.launch"
+
+
